Moved the worker based DrawingSystem into the DrawingSystem.js file.
