+++ conway_module/lib/core/DrawingSystem.js
@@ -1,58 +1,127 @@
-const {CanvasBasedSystem} = require('./System.js')
-const DefaultConfig = require('./DefaultConfig.js')
+
 const {DrawingStateManager} = require('./DrawingStateManager.js');
+const SceneManager = require('./SceneManager.js');
+
 
 /**
- * A system for drawing on a grid.
- * @extends CanvasBasedSystem
+ * The possible states the drawing system can be in.
+ * @private
  */
-class DrawingSystem extends CanvasBasedSystem{
+const States = {
+  IDLE: 'idle',
+  UPDATING: 'updating'
+};
+
+/**
+ * A system for drawing on a grid. Designed to be used in the context of a web worker.
+*/
+class DrawingSystem {
 	/**
 	 * Creates a new drawing system.
-	 * @param {Window} window - The DOM's window object.
-	 * @param {HTMLCanvasContext} htmlCanvasContext - An HTML5 Canvas 2D context.
-	 * @param {object} config - A configuration object.
 	 */
-	constructor(window, htmlCanvasContext, config = DefaultConfig){
-		super(window, htmlCanvasContext, config)
-		this.stateManager = new DrawingStateManager(this.config)
-	}
+  constructor() {
+    this.config = {
+      zoom: 1
+    };
+    this.stateManager = new DrawingStateManager(this.config);
+    this.scene = new SceneManager();
+    this.displayStorageStructure = false;
+    this.state = States.IDLE
+  }
 
 	/**
-	 * Override parent
-	 * @private
+	 * @returns {State} Returns the active state.
 	 */
-	getStateManager(){
-		return this.stateManager
-	}
+  getState() {
+    return this.state;
+  }
 
 	/**
-	 * Override parent
-	 * @private
+	 * Replaces the current configuration.
+	 * @param {*} config 
 	 */
-	update(frame){
-		let updateTime = 'Drawing System Update Time';
-		console.time(updateTime);
-		this.getStateManager().stageStorage(this.scene, this.displayStorageStructure)
-		this.getStateManager().processCells(this.scene)
-		console.timeEnd(updateTime);
+  setConfig(config) {
+    this.config = config;
+    this.stateManager.setConfig(this.config);
+  }
+
+	/**
+	 * @returns {SceneManager} The current scene.
+	 */
+  getScene() {
+    return this.scene;
+  }
+
+	/**
+	 * @returns {DrawingStateManager} Returns the state manager for the drawing system.
+	 */
+  getStateManager() {
+    return this.stateManager
+  }
+
+  update() {
+    this.state = States.UPDATING
+    this.scene.clear();
+    this.getStateManager().stageStorage(this.scene, this.displayStorageStructure);
+    this.getStateManager().processCells(this.scene);
+    this.state = States.IDLE;
+  }
+
+	canUpdate(){
+		return this.state === States.IDLE;
 	}
 
 	/**
 	 * Used to preload the drawing system with alive cells.
 	 * @param {Cell[]} cells - An array of alive cells.
 	 */
-	setCells(cells){
-		this.getStateManager().setCells(cells)
-	}
+  setCells(cells) {
+    this.getStateManager().setCells(cells)
+  }
+
+  /**
+	 * Provides a deep copy of the currently alive cells.
+	 * @returns {Cell[]}
+	 */
+  getCells() {
+    return this.getStateManager().getCells()
+  }
+
+  /**
+	 * Sets the cell size to use.
+	 * @param {number} size
+	 */
+  setCellSize(size) {
+    this.config.zoom = size
+  }
 
 	/**
 	 * Flips a grid cell to alive or dead.
 	 * @param {number} x - The X coordinate of the cell.
 	 * @param {number} y - The Y coordinate of the cell.
 	 */
-	toggleCell(x,y){
-		this.getStateManager().toggleCell(x,y)
+  toggleCell(x, y) {
+    if (this.state === States.IDLE) {
+      this.state = States.UPDATING
+      this.getStateManager().toggleCell(x, y);
+      this.state = States.IDLE;
+    }
+  }
+
+  /**
+	 * Sets whether to draw the quad tree.
+	 * @param {boolean} display
+	 */
+  displayStorage(display) {
+    this.displayStorageStructure = display
+  }
+
+  /**
+	 * Clears the simulation.
+	 */
+	reset(){
+		this.scene.clear()
+		this.getStateManager().clear()
 	}
 }
